Derives a single or ensemble of bivariate design events

Calculates the isoline and relative probability of events on the isoline, where the data contains events from two populations. Outputs the single "most-likely" design event or an ensemble of possible design events obtained by sampling along the isoline according to these relative probabilities. The design event under the assumption of full dependence is also computed. Isoline is derived by calculating annual exceedance probabilities from both copula models on a user-specified grid rather by overlaying the partial isolines from the two copula models as in Design_Event_2D.

Usage

Design_Event_2D_Multi_Pop(
  Data,
  Data_Con1,
  Data_Con2,
  Data_Con3,
  Data_Con4,
  u1,
  u2,
  u3,
  u4,
  Thres1 = NA,
  Thres2 = NA,
  Thres3 = NA,
  Thres4 = NA,
  N_Both_1,
  N_Both_2,
  Copula_Family1,
  Copula_Family2,
  Copula_Family3,
  Copula_Family4,
  Marginal_Dist1,
  Marginal_Dist2,
  Marginal_Dist3,
  Marginal_Dist4,
  Marginal_Dist1_Par = NA,
  Marginal_Dist2_Par = NA,
  Marginal_Dist3_Par = NA,
  Marginal_Dist4_Par = NA,
  Con1 = "Rainfall",
  Con2 = "OsWL",
  Con3 = "Rainfall",
  Con4 = "OsWL",
  GPD1 = NA,
  GPD2 = NA,
  GPD3 = NA,
  GPD4 = NA,
  Rate_Con1 = NA,
  Rate_Con2 = NA,
  Rate_Con3 = NA,
  Rate_Con4 = NA,
  Tab1 = NULL,
  Tab2 = NULL,
  Tab3 = NULL,
  Tab4 = NULL,
  mu = 365.25,
  GPD_Bayes = FALSE,
  Decimal_Place = 2,
  Grid_x_min = NA,
  Grid_x_max = NA,
  Grid_y_min = NA,
  Grid_y_max = NA,
  Grid_x_interval = NA,
  Grid_y_interval = NA,
  RP,
  x_lab = "Rainfall (mm)",
  y_lab = "O-sWL (mNGVD 29)",
  x_lim_min = NA,
  x_lim_max = NA,
  y_lim_min = NA,
  y_lim_max = NA,
  Isoline_Probs = "Sample",
  N = 10^6,
  N_Ensemble = 0,
  Sim_Max = 10,
  Plot_Quantile_Isoline = FALSE
)

Arguments

Data

Data frame of dimension nx2 containing two co-occurring time series of length n.

Data_Con1

Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column for population 1.

Data_Con2

Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column for population 1.

Data_Con3

Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the first column for population 2.

Data_Con4

Data frame containing the conditional sample (declustered excesses paired with concurrent values of other variable), conditioned on the variable in the second column for population 2.

u1

Numeric vector of length one specifying the threshold, expressed as a quantile, above which the variable in the first column was sampled in Data_Con1.

u2

Numeric vector of length one specifying the threshold, expressed as a quantile, above which the variable in the second column was sampled in Data_Con2.

u3

Numeric vector of length one specifying the threshold, expressed as a quantile, above which the variable in the first column was sampled in Data_Con3.

u4

Numeric vector of length one specifying the threshold, expressed as a quantile, above which the variable in the second column was sampled in Data_Con4.

Thres1

Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in Data_Con1. Only one of u1 and Thres1 should be supplied. Default is NA.

Thres2

Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in Data_Con2. Only one of u2 and Thres2 should be supplied. Default is NA.

Thres3

Numeric vector of length one specifying the threshold above which the variable in the first column was sampled in Data_Con3. Only one of u3 and Thres3 should be supplied. Default is NA.

Thres4

Numeric vector of length one specifying the threshold above which the variable in the second column was sampled in Data_Con4. Only one of u4 and Thres4 should be supplied. Default is NA.

N_Both_1

Numeric vector of length one specifying the number of data points in population 1 that feature in both conditional samples.

N_Both_2

Numeric vector of length one specifying the number of data points in population 1 that feature in both conditional samples.

Copula_Family1

Numeric vector of length one specifying the copula family used to model the Data_Con1 dataset.

Copula_Family2

Numeric vector of length one specifying the copula family used to model the Data_Con2 dataset.

Copula_Family3

Numeric vector of length one specifying the copula family used to model the Data_Con3 dataset.

Copula_Family4

Numeric vector of length one specifying the copula family used to model the Data_Con4 dataset.

Marginal_Dist1

Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in Data_Con1.

Marginal_Dist2

Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in Data_Con2.

Marginal_Dist3

Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in Data_Con3.

Marginal_Dist4

Character vector of length one specifying (non-extreme) distribution used to model the marginal distribution of the non-conditioned variable in Data_Con4.

Marginal_Dist1_Par

Character vector of length one specifying (non-extreme) parameters of Marginal_Dist1. Default is NA, specified distribution is fit within the procedure.

Marginal_Dist2_Par

Character vector of length one specifying (non-extreme) parameters of Marginal_Dist2. Default is NA, specified distribution is fit within the procedure.

Marginal_Dist3_Par

Character vector of length one specifying (non-extreme) parameters of Marginal_Dist3. Default is NA, specified distribution is fit within the procedure.

Marginal_Dist4_Par

Character vector of length one specifying (non-extreme) parameters of Marginal_Dist4. Default is NA, specified distribution is fit within the procedure.

Con1

Character vector of length one specifying the name of variable in the first column of Data_Con1.

Con2

Character vector of length one specifying the name of variable in the second column of Data_Con2.

Con3

Character vector of length one specifying the name of variable in the first column of Data_Con3.

Con4

Character vector of length one specifying the name of variable in the second column of Data_Con4.

GPD1

Output of GPD_Fit applied to variable con1 i.e., GPD fit con1. Default NULL. Only one of u1, Thres1, GPD1 and Tab1 is required.

GPD2

Output of GPD_Fit applied to variable con2 i.e., GPD fit con2. Default NULL. Only one of u2, Thres2, GPD2 and Tab2 is required.

GPD3

Output of GPD_Fit applied to variable con3 i.e., GPD fit con3. Default NULL. Only one of u3, Thres3, GPD3 and Tab3 is required.

GPD4

Output of GPD_Fit applied to variable con4 i.e., GPD fit con4. Default NULL. Only one of u4, Thres4, GPD4 and Tab4 is required.

Rate_Con1

Numeric vector of length one specifying the occurrence rate of observations in Data_Con1. Default is NA.

Rate_Con2

Numeric vector of length one specifying the occurrence rate of observations in Data_Con2. Default is NA.

Rate_Con3

Numeric vector of length one specifying the occurrence rate of observations in Data_Con3. Default is NA.

Rate_Con4

Numeric vector of length one specifying the occurrence rate of observations in Data_Con4. Default is NA.

Tab1

Data frame specifying the return periods of variable con1, when conditioning on con1. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of con1 for the inter-arrival time (1/rate) of the sample. Default is NULL.

Tab2

Data frame specifying the return periods of variable con2, when conditioning on con2. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of con2 for the inter-arrival time (1/rate) of the sample. Default is NULL.

Tab3

Data frame specifying the return periods of variable con3, when conditioning on con3. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of con3 for the inter-arrival time (1/rate) of the sample. Default is NULL.

Tab4

Data frame specifying the return periods of variable con4, when conditioning on con4. First column specifies the return period and the second column gives the corresponding levels. First row must contain the return level of con4 for the inter-arrival time (1/rate) of the sample. Default is NULL.

mu

Numeric vector of length one specifying the (average) occurrence frequency of events in Data. Default is 365.25, daily data.

GPD_Bayes

Logical; indicating whether to use a Bayesian approach to estimate GPD parameters. This involves applying a penalty to the likelihood to aid in the stability of the optimization procedure. Default is FALSE.

Decimal_Place

Numeric vector specifying the number of decimal places to which to specify the isoline. Default is 2

Grid_x_min

Numeric vector of length one specifying the minimum value of the variable in first column of Data contained in the grid.

Grid_x_max

Numeric vector of length one specifying the maximum value of the variable in first column of Data contained in the grid.

Grid_y_min

Numeric vector of length one specifying the minimum value of the variable in second column of Data contained in the grid.

Grid_y_max

Numeric vector of length one specifying the maximum value of the variable in second column of Data contained in the grid.

Grid_x_interval

Numeric vector of length one specifying the resolution of the grid in terms of the variable in first column of Date. Default is an interval 2 of between consecutive values.

Grid_y_interval

Numeric vector of length one specifying the resolution of the grid in terms of the variable in second column of Date. Default is an interval 0.1 of between consecutive values.

RP

Numeric vector specifying the return periods of interest.

x_lab

Character vector specifying the x-axis label.

y_lab

Character vector specifying the y-axis label.

x_lim_min

Numeric vector of length one specifying x-axis minimum. Default is NA.

x_lim_max

Numeric vector of length one specifying x-axis maximum. Default is NA.

y_lim_min

Numeric vector of length one specifying y-axis minimum. Default is NA.

y_lim_max

Numeric vector of length one specifying y-axis maximum. Default is NA.

Isoline_Probs

Character vector of length one specifying whether to calculate relative probabilities of points on the isoline from a "Sample" simulated from the fitted copula models or from the "Observations".Default is "Sample".

N

Numeric vector of length one specifying the size of the sample from the fitted joint distributions used to estimate the density along an isoline. Samples are collected from the two joint distribution with proportions consistent with the total number of extreme events conditioned on each variable. Default is 10^6

N_Ensemble

Numeric vector of length one specifying the number of possible design events sampled along the isoline of interest.

Sim_Max

Numeric vector of length one specifying the maximum value, given as a multiple of the largest observation of each variable, permitted in the sample used to estimate the (relative) probabilities along the isoline.

Plot_Quantile_Isoline

Logical; indicating whether to first plot the quantile isoline. Default is FALSE.

Value

Plot of all the observations (grey circles) as well as the declustered excesses above Thres1 (blue circles) or Thres2 (blue circles), observations may belong to both conditional samples. Also shown is the isoline associated with RP contoured according to their relative probability of occurrence on the basis of the sample from the two joint distributions, the "most likely" design event (black diamond), and design event under the assumption of full dependence (black triangle) are also shown in the plot. The function also returns a list comprising the design events assuming full dependence "FullDependence", as well as once the dependence between the variables is accounted for the "Most likley" "MostLikelyEvent" as well as an "Ensemble" of possible design events and relative probabilities of events on the isoline Contour. The quantile isolines with Quantile_Isoline_1 and Quantile_Isoline_2, and GPD thresholds with Threshold_1 and Threshold_2.

See also