HURDAT

Appends a data frame with the names of any storms in the HURDAT database with a center of circulation within a specified radius of a location.

Usage

HURDAT(Data, lat.loc, lon.loc, rad)

Arguments

Data

A data frame where the first column is a date or date_time object.

lat.loc

Numeric vector of length one specifying the latitude (in degrees)of the location of interest.

lon.loc

Numeric vector of length one specifying the longitude (in degrees) of the location of interest.

rad

Vector of length one, specifying radius (in km) about the location of interest to report storm names.

Value

The data frame Data with an additional column containing the named storms.