MultiHazard: Tools for modeling compound events

The `MultiHazard` package provides tools for stationary multivariate statistical modeling, for example, to estimate the joint distribution of MULTIple co-occurring HAZARDs. The package contains functions for pre-processing data including imputing missing values, detrending and declustering time series as well as analyzing pairwise correlations over a range of lags. Functionality is also built in to impliment the conditional sampling - copula theory approach in Jane et al. (2020) including the automated threshold selection approach in Solari et al. (2017). Tools are provided for selecting the best fitting amongst an array of (non-extreme, truncated and non-truncated) parametric marginal distributions, and, copulas to model the dependence structure. The package contains a function that calculates joint probability contours using the method of overlaying (conditional) contours given in Bender et al. (2016), and extracting design events such as the 'most likely' event or an ensemble of possible design events. The package also provides the capability of fitting and simulating synthetic records from three higher dimensional approaches - standard (elliptic/Archimedean) copulas, Pair Copula Constructions (PCCs) and the conditional threshold exceedance approach of Heffernan and Tawn (2004). Finally, a function that calculates the time for a user-specified height of sea level rise to occur under various scenarios is supplied.

See also

Author

Maintainer: Robert Jane r.jane@ucf.edu (ORCID)

Other contributors:

Thomas Wahl t.wahl@ucf.edu (ORCID) [contributor]

Francisco Pena fpena@sfwmd.gov (ORCID) [contributor]

Jayantha Obeysekera jobeysek@fiu.edu (ORCID) [contributor]

Callum Murphy-Barltrop callum.murphy-barltrop@tu-dresden.de (ORCID) [contributor]

Javed Ali javed.ali@ucf.edu (ORCID) [contributor]

Pravin Maduwantha pravin@ucf.edu (ORCID) [contributor]

Huazhi Li huazhi.li@vu.nl (ORCID) [contributor]

Victor Malagon Santos victor.malagon.santos@nioz.nlu (ORCID) [contributor]