NOAA sea-level rise scenarios

Time (in years) for a specified amount of sea-level rise (SLR) to occur at Miami Beach according to the five SLR scenarios in NOAA 2017 report titled "Global and Regional Sea Level Rise Scenarios for the United States".

NOAA_SLR(
  OsWL_req,
  SLR_scen = c("High", "Intermediate", "Low"),
  Input_unit = "m",
  Year.Inital = 2020
)

Arguments

OsWL_req

Numeric vector of SLR required.

SLR_scen

Character vector specifying which of the NOAA (2017) scenarios to consider. Options include High, Intermediate high Int.High, Intermediate, Intermediate low (Int.Low) and Low.

Input_unit

Character vector of length one; specifying units of SLR. Default is meters "m", other option is feet "ft".

Year.Inital

Character vector of length one; specifying the current year.

Value

List comprising the specified Threshold as the quantile of the conditioning variable above which declustered excesses are paired with co-occurrences of the other variable, the resulting two-dimensional sample data and name of the conditioning variable.

Examples

NOAA_SLR(OsWL_req=seq(0,1,0.01),SLR_scen = c("High","Intermediate","Low"),
                   Input_unit="m")
#> Error in NOAA_SLR(OsWL_req = seq(0, 1, 0.01), SLR_scen = c("High", "Intermediate",     "Low"), Input_unit = "m"): object 'NOAA2017' not found