Declustered Time Series for Case Study Site S-22 in Jane et al. (2020)

Time series of rainfall totals from the Miami International Airport gauge, ocean-side water levels at control structure S-22, and groundwater at Well G-580A (with missing values interpolated using groundwater levels at Well G-860). All time series were declustered using a peaks-over-threshold approach (runs method with a 0.98 quantile threshold and a 3-day separation criterion). Ocean-side water levels and groundwater levels were each detrended using a 3-month moving average window prior to declustering.

data(S22.Detrend.Declustered.df)

Format

A data frame with XX rows and 4 variables:

Date

Date of observation (YYYY-MM-DD)

Rainfall

Rainfall totals (inches)

OsWL

Ocean-side water levels (ft NGVD29)

Groundwater

Groundwater levels (ft NGVD29)

Source

Rainfall totals from the NOAA National Centers for Environmental Information: https://www.ncdc.noaa.gov/ Ocean-side water levels and groundwater levels from the South Florida Water Management District: https://www.sfwmd.gov/science-data/dbhydro/

References

Jane, R., Cadavid, L., Obeysekera, J., and Wahl, T.: Multivariate statistical modelling of the drivers of compound flood events in south Florida, Nat. Hazards Earth Syst. Sci., 20, 2681–2699. https://doi.org/10.5194/nhess-20-2681-2020