Fit an Archimedean/elliptic copula model - Fit

Fit a n-dimensional Archimedean or elliptic copula model. Function is simply a repackaging of the fitCopula function in the copula package.

Usage

Standard_Copula_Fit(Data, Copula_Type = "Gaussian")

Arguments

Data

Data frame containing n at least partially concurrent time series. First column may be a "Date" object. Can be Dataframe_Combine output.

Copula_Type

Type of elliptical copula to be fitted, options are "Gaussian" (Default), "tcopula", "Gumbel", "Clayton" and "Frank".

Value

List comprising the Copula_Type and the fitted copula Model object.

See also

Examples

cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Gaussian")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="tcopula")
#> Warning: the covariance matrix of the parameter estimates is computed as if 'df.fixed = TRUE' with df = 16.3245687783949
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Gumbel")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Clayton")
cop<-Standard_Copula_Fit(Data=S20.Detrend.df,Copula_Type="Frank")