Selecting best fitting standard (elliptical and Archimedean) copula

Fits five n-dimensional standard copula to a dataset and returns their corresponding AIC values.

Usage

Standard_Copula_Sel(Data)

Arguments

Data

Data frame containing n at least partially concurrent time series, detrended if necessary. Time steps must be equally spaced, with missing values assigned NA. First object may be a "Date" object. Can be Dataframe_Combine output.

Value

Data frame containing copula name in column 1 and associated AIC in column 2. Parameters are estimated using the fitCopula() function in copula package using maximum pseudo-likelihood estimator "mpl". See fitCopula for a more thorough explanation.

See also

Examples

Standard_Copula_Sel(Data=S20.Detrend.df)
#> Warning: the covariance matrix of the parameter estimates is computed as if 'df.fixed = TRUE' with df = 16.3245687783949
#>     Copula        AIC
#> 1 Gaussian -1397.8675
#> 2    t-cop -1442.5394
#> 3   Gumbel  -884.8532
#> 4  Clayton  -572.0618
#> 5    Frank  -865.2858