Rainfall and O-sWL time series plots

Plots a user specified number of synthetic events where at least O-sWL or rainfall peak exceeds a high threshold. .

Usage

Time_Series_Plot(
  Rainfall_Series,
  Oswl_Time_Series,
  Sample,
  Con_Variable,
  Buffer = 6,
  Intensity = NA,
  Event_ID = 1:16,
  Row = 4,
  Col = 4,
  Mar = c(4.2, 4.5, 1.5, 3.5)
)

Arguments

Rainfall_Series

Data frame with rows comprising time series of rainfall totals associated with cluster maximum of the rainfall series.

Oswl_Time_Series

Data frame with rows comprising the water level curves associated with the simulated events in Sample.

Sample

Data frame containing the simulated events. Columns (and their names) required by the function are rainfall peak (Rainfall), O-sWL peak (OsWL), their lag time (Lag), and the ID of the sampled rainfall event (samp).

Con_Variable

Character vector of length one specifying the conditioning variable of the events in Sample.

Buffer

Numeric vector of length one specifying the extension of the x-axis before and after the rainfall event when Con_Variable == "Rainfall". Default is 6.

Intensity

Numeric vector specifying the "intensity" of the O-sWL events in Sample. Default is NA.

Event_ID

Numeric vector specifying the events in Sample to be plot.

Row

Numeric vector of length one specifying the number of rows of subplots in the Figure.

Col

Numeric vector of length one specifying the number of columns of subplots in the Figure. Product of Row and Col must be equal to or greater than Event_ID.

Mar

Numeric vector of length one specifying the margin at the (bottom,left,top,right) of the subplots. Default is c(4.2,4.5,1.5,3.5).

Value

Figure containing a (Row * Col) matrix of subplots each displaying the hyetogaph (grey bars) and water level curve (blue lines) comprising an event.

See also