Generate annual maximum series

Extract annual maximum in years with over a user-defined proportion of non-missing values.

Usage

Annual_Max(Data_Detrend, Complete_Prop = 0.8)

Arguments

Data_Detrend

Data frame containing two columns. In column:

1 A "Date" object of equally spaced discrete time steps.

2 Numeric vector containing corresponding time series values.

Complete_Prop

Minimum proportion of non-missing values in an annual record for the annual maximum to be extracted. Default is 0.8.

Value

List comprising the index of the annual maximum Event and the annual maximum values AM.

Examples

#Converting
S20_T_MAX_Daily_Completed_Detrend_Declustered[,1] <-
                  as.Date(S20_T_MAX_Daily_Completed_Detrend_Declustered[,1], format = "%m/%d/%Y")
#Finding annual maximum of detrended time series
Annual_Max(Data_Detrend=S20_T_MAX_Daily_Completed_Detrend_Declustered[,c(1,4)])
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> $Event
#>  [1]   220   464   962  1298  1675  2023  2300  2708  3107  3511  3878  4216
#> [13]  4570  4977  5335  5725  6021  6419  6796  7085  7455  7895  8234  8498
#> [25]  8999  9367  9586 10099 10542 10777 11162 11516 11914 12147 12650 13015
#> [37] 13329 13709 14070 14431 14839 15164 15538 15921 16306 16651 17011 17370
#> [49] 17702 18060 18267
#> 
#> $AM
#>  [1] 2.540000 2.140000 1.978000 1.840000 2.200000 2.470000 2.220000 2.222000
#>  [9] 1.959000 2.320000 2.066000 2.090000 2.683000 2.220000 2.060000 2.212000
#> [17] 2.200000 2.100000 2.370000 2.260000 1.950000 2.050000 2.530000 2.705000
#> [25] 2.090327 3.284000 2.269000 3.016000 2.726000 2.854000 3.323000 2.967000
#> [33] 2.620000 2.400000 2.450000 2.255978 3.346512 2.221000 2.916000 2.510000
#> [41] 2.360000 2.760000 2.710000 2.673000 2.287000 2.359000 2.917000 2.772000
#> [49] 5.854000 2.181000 1.883000
#>