Fits a single generalized Pareto distribution - Fit

Fit a Generalized Pareto Distribution (GPD) to a declustered dataset.

GPD_Fit(
  Data,
  Data_Full,
  u = 0.95,
  Thres = NA,
  mu = 365.25,
  GPD_Bayes = TRUE,
  Method = "Standard",
  min.RI = 1,
  max.RI = 100,
  PLOT = FALSE,
  xlab_hist = "Data",
  y_lab = "Data"
)

Arguments

Data

Numeric vector containing the declusted data.

Data_Full

Numeric vector containing the non-declustered data.

u

GPD threshold expressed as a quantile [0,1] of Data vector. Default is 0.95.

Thres

GPD threshold expressed on the original scale of the "Data". Only one of u and Thres should be supplied. Default is NA.

mu

Numeric vector of length one specifying (average) occurrence frequency of events in the Data_Full input. Default is 365.25.

GPD_Bayes

Logical; indicating whether to use a Bayesian approach to estimate GPD parameters. This involves applying a penalty to the likelihood to aid in the stability of the optimization procedure. Default is TRUE.

Method

Character vector of length one specifying the method of choosing the threshold. "Standard" (default) chooses the exact threshold specified as either "u" or "th", whereas "Solari" selects the minimum exceedence of the "Data" above the user-specified threshold.

min.RI

Numeric vector of length one specifying the minimum return period in the return level plot. Default is 1.

max.RI

Numeric vector of length one specifying the maximum return period in the return level plot. Default is 100.

PLOT

Logical; indicating whether to plot diagnostics. Default is FALSE.

xlab_hist

Character vector of length one. Histogram x-axis label. Default is "Data".

y_lab

Character vector of length one. Histogram x-axis label. Default is "Data".

Value

List comprising the GPD Threshold, shape parameter xi and scale parameters sigma along with their standard errors sigma.SE and xi.SE.

Details

For excesses of a variable X over a suitably high threshold u the fitted GPD model is parameterized as follows: $$P( X > x| X > u) = \left[1 + \xi \frac{(x-u)}{\sigma}\right]^{-\frac{1}{\xi}}_{+}$$ where \(\xi\) and \(\sigma>0\) are the shape and scale parameters of the GPD and \([y]_{+}=max(y,0)\).

Examples

#Decluster time series
S20T_decl = Decluster(Data=S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
#Fit GPD
GPD_Fit(Data=S20T_decl$Declustered,
        Data_Full=S20_T_MAX_Daily_Completed_Detrend_Declustered$Detrend)
#> $Threshold
#> [1] 1.823222
#> 
#> $Rate
#> [1] 4.74506
#> 
#> $sigma
#> [1] 0.1594508
#> 
#> $xi
#> [1] 0.172105
#> 
#> $sigma.SE
#> [1] 0.09144518
#> 
#> $xi.SE
#> [1] 0.06507229
#>